ubx: updates for protocol 27 & M8P
Sorry for the large commit :/

Deprecated messages:
- CFG-RRT
- CFG-MSG
- CFG-RATE
- CFG-CFG
- CFG-NAV5
- CFG-TMODE3
- MON-HW

With the new messages:
- CFG-VAL{GET,SET,DEL}
- MON-RF

So this affects a large part of the driver, including auto-config.
The auto-detection now first tries to use the new config command,
and based on the result of that we know if we can use protocol version 27
or higher.

Other changes include:
- add board detection
- higher baudrate for M8* and F9* devices (115200 instead of 38400).
  Reduces latency.
- 10Hz update rate for F9*

Would be nice to have:
- detect on which port we are running (there is no message for that). At
  the moment we configure UART1 and USB