sbf: fix overrun on invalid length
Before checking the CRC across header and payload, we need to check if
the length is valid. If the length is longer than what we have read into
the buffer, it must be invalid and we don't have to bother with the CRC.

This should fix a potential segfault where the CRC overruns the
allocated buffer due to a corrupt length field.