ubx: do not warn on invalid len
Handling this as a warning is too sensitive.

It only requires the msg id byte or one of the len bytes to get a bit flipped due to EMI
and the log can get spammed with these messages in a noisy environment.

The parser does handle these situations without any problems. In case too many messages
are lost the timeout will kick in anyway